@@ -1,14 +1,14 @@
 name: lexpro
 description: Dart lexer for various programming languages base on lex v0.0.1 originally published by thlorenz10@gmail.com edit by jackyanjiaqi@gmail.com
-version: 1.1.0+1
+version: 2.0.0
 homepage: https://www.zhihu.com/people/digitalspritejack
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  equatable: ^1.2.5
+  equatable: ^2.0.0
 
 dev_dependencies:
-  test: ^1.15.3
+  test: ^1.16.0
 
